Fix a tests/math bug
This was a particularly pernicious bug. Ultimately, the fix is here:

```
-    is_const_or_var && subst[&x] != subst[&c]
+    is_const_or_var && egraph.find(subst[&x]) != egraph.find(subst[&c])
```

We weren't canonicalizing the ids in `c_is_const_or_var_and_not_x`,
which causes both d-variable and d-constant to fire at the same time,
placing 0 and 1 in the same eclass.

Initially, I thought this was due to divide-by-zero in the rules, and I
left in all the checks and assertions that led me to the true bug.